[MOBDW-233] Refactor/send receive page (#115)
* Refactor Send/Receive MOB page

* Decided not to separate modal

* Small type details

* Story for ReceiveMob - it won't build, it crashes

* Don't comment mkdir
