+++ app/pages/SendReceive/ReceiveMob.view/index.ts
@@ -1,0 +1,4 @@
+import { ReceiveMob } from './ReceiveMob.view';
+
+export default ReceiveMob;
+export { ReceiveMob };
