+++ app/pages/index.ts
@@ -1,4 +1,5 @@
 import { ContactsPage } from './Contacts';
 import { HistoryPage } from './History';
+import { SendReceivePage } from './SendReceive';
 
-export { ContactsPage, HistoryPage };
+export { ContactsPage, SendReceivePage, HistoryPage };
