--- app/views/wallet/TransactionView/TransactionView.stories.tsx
@@ -1,27 +1,0 @@
-// import React, { ComponentProps } from 'react';
-
-// import { Story } from '@storybook/react';
-// import { SnackbarProvider } from 'notistack';
-
-// import TransactionView from './TransactionView';
-
-// export default {
-//   component: TransactionView,
-//   title: 'History/Transaction/View',
-// };
-
-// const Template: Story<ComponentProps<typeof TransactionView>> = (args) => (
-//   <TransactionView {...args} />
-// );
-
-// export const Primary = Template.bind({});
-
-// Primary.decorators = [
-//   (StoryComponent) => (
-//     <SnackbarProvider>
-//       <StoryComponent />
-//     </SnackbarProvider>
-//   ),
-// ];
-
-// Primary.args = {};
