--- app/views/wallet/TransactionView/index.ts
@@ -1,3 +1,0 @@
-import TransactionView from './TransactionView';
-
-export default TransactionView;
